refactor(routing/http): preserve http-path in ToMultiaddr, extract interop helpers
- ToMultiaddr() now appends /http-path/ when the URL has a non-empty path,
  matching go-multiaddr http-path protocol conventions
- moved peerIDFromDIDKey, peerIDPlaceholderFromArbitraryID, hasHTTPURL
  to contentrouter/interop.go for better organization
- added unit tests for hasHTTPURL covering http, https, non-HTTP,
  multiaddr, mixed, empty, and invalid address cases
- fixed staticcheck lint suppression for SchemaBitswap in ndjson test
